Lead_Identification/detection/agent_tavily/backend_1_enrichment/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
tavily-python
beautifulsoup4
requests
faiss-cpu
langchain
langchain-google-genai
langgraph
langchain-community
pydantic-settings
